lark/home/crawler/requirements.txt
beautifulsoup4==4.12.2
lxml==4.9.3
requests==2.31.0
python-dotenv==1.0.0
lark-oapi==1.0.0
openai==1.3.7
larksuiteoapi==1.0.0 